Break up metadata
